\name{ce.IDsMap}
\alias{ce.IDsMap}
\docType{data}
\title{
  map file for converting gene name or sequence name of \emph{Caenorhabditis elegans} to wormbase identifier
}
\description{
  map file for converting gene name or sequence name of \emph{Caenorhabditis elegans} to wormbase identifier
}
\usage{data(ce.IDsMap)}
\format{
   character vector
}
\details{
	character vecotr with gene name or sequence name as names and wormbase identifier as values.
}
\examples{
data(ce.IDsMap)
head(ce.IDsMap)
}
\keyword{datasets}
