\name{ce.interactionmap}
\alias{ce.interactionmap}
\docType{data}
\title{
  transcript regulatory map of \emph{Caenorhabditis elegans}
}
\description{
  transcript regulatory map of \emph{Caenorhabditis elegans}
}
\usage{data(ce.interactionmap)}
\format{
   dataframe
}
\details{
	transcript regulatory map of \emph{Caenorhabditis elegans} is generated using databases edgedb and microCosm Targets.
}
\examples{
data(ce.interactionmap)
head(ce.interactionmap)
}
\keyword{datasets}
