\name{ce.mapIDs}
\alias{ce.mapIDs}
\docType{data}
\title{
  map file for converting from wormbase identifier to \emph{Caenorhabditis elegans} gene name
}
\description{
  map file for converting from wormbase identifier to \emph{Caenorhabditis elegans} gene name
}
\usage{data(ce.mapIDs)}
\format{
   character vector
}
\details{
	character vecotr with wormbase identifier as names and gene name as values.
}
\examples{
data(ce.mapIDs)
head(ce.mapIDs)
}
\keyword{datasets}
