\name{ce.miRNA.map}
\alias{ce.miRNA.map}
\docType{data}
\title{
  micro RNA of \emph{Caenorhabditis elegans}
}
\description{
  micro RNA of \emph{Caenorhabditis elegans}
}
\usage{data(ce.miRNA.map)}
\format{
   dataframe
}
\details{
	The first column is wormbase identifier. And the second column is miRNA names.
}
\examples{
data(ce.miRNA.map)
head(ce.miRNA.map)
}
\keyword{datasets}
