% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNetworkBuilder.R
\docType{package}
\name{GeneNetworkBuilder-package}
\alias{GeneNetworkBuilder}
\alias{GeneNetworkBuilder-package}
\title{Build Regulatory Network from ChIP-chip/ChIP-seq and Expression Data}
\description{
Appliation for discovering direct or indirect targets of
transcription factors using ChIP-chip or ChIP-seq, and
microarray or RNA-seq gene expression data. Inputting a list of
genes of potential targets of one TF from ChIP-chip or
ChIP-seq, and the gene expression results, GeneNetworkBuilder
generates a regulatory network of the TF.
}
\author{
\strong{Maintainer}: Jianhong Ou Developer \email{jou@morgridge.org}

Authors:
\itemize{
  \item Lihua Julie Zhu Developer \email{Julie.Zhu@umassmed.edu}
}

}
