% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{hs.IDsMap}
\alias{hs.IDsMap}
\title{map file for converting gene name or sequence name of \emph{Homo sapiens} to Entrez identifier}
\format{character vector}
\usage{
hs.IDsMap
}
\description{
map file for converting gene name or sequence name of \emph{Homo sapiens} to Entrez identifier
}
\details{
character vecotr with gene name as names and Entrez identifier as values.
}
\examples{
data(hs.IDsMap)
head(hs.IDsMap)

}
\keyword{datasets}
