\name{hs.interactionmap}
\alias{hs.interactionmap}
\docType{data}
\title{
  transcript regulation map of \emph{Homo sapiens}
}
\description{
  transcript regulation map of \emph{Homo sapiens}
}
\usage{data(hs.interactionmap)}
\format{
   dataframe
}
\details{
	transcript regulatory map of \emph{Homo sapiens} is generated using databases FANTOM, 
mirGen and microCosm Targets.
}
\examples{
data(hs.interactionmap)
head(hs.interactionmap)
}
\keyword{datasets}
