\name{hs.mapIDs}
\alias{hs.mapIDs}
\docType{data}
\title{
  map file for converting from Entrez identifier to \emph{Homo sapiens} gene name
}
\description{
  map file for converting from Entrez identifier to \emph{Homo sapiens} gene name
}
\usage{data(hs.mapIDs)}
\format{
   character vector
}
\details{
	character vecotr with Entrez identifier as names and gene name as values.
}
\examples{
data(hs.mapIDs)
head(hs.mapIDs)
}
\keyword{datasets}
