\name{hs.miRNA.map}
\alias{hs.miRNA.map}
\docType{data}
\title{
  micro RNA of \emph{Homo sapiens}
}
\description{
  micro RNA of \emph{Homo sapiens}
}
\usage{data(hs.miRNA.map)}
\format{
   dataframe
}
\details{
	The first column is entrez identifier. And the second column is miRNA names.
}
\examples{
data(hs.miRNA.map)
head(hs.miRNA.map)
}
\keyword{datasets}
