% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportNetwork.R
\name{saveXGMML}
\alias{saveXGMML}
\title{Save network as xgmml}
\usage{
saveXGMML(network, file, ...)
}
\arguments{
\item{network}{output of \link{browseNetwork}}

\item{file}{Name of the file to save to.}

\item{...}{Parameter could be used by saveXML}
}
\description{
Save graph into xgmml format.
}
\keyword{IO}
