% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetNetwork.R
\name{subsetNetwork}
\alias{subsetNetwork}
\title{Subset a polished network}
\usage{
subsetNetwork(graph, genes)
}
\arguments{
\item{graph}{A graphNEL object. The output of polishNetwork.}

\item{genes}{A list of nodes names}
}
\value{
An object of graph.
}
\description{
Subset the output of polishNetwork by a list of nodes name
}
\examples{
library(graph)
set.seed(123)
g1 <- randomEGraph(LETTERS[seq.int(15)], edges=100)
g1 <- subsetNetwork(g1, LETTERS[seq.int(5)])
plot(g1)
}
