% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ce.IDsMap}
\alias{ce.IDsMap}
\title{C.elegns gene name to wormbase identifier map}
\format{
character vector
}
\source{
\url{http://www.wormbase.org/}
}
\usage{
ce.IDsMap
}
\description{
map file for converting gene name or sequence name of 
\emph{Caenorhabditis elegans} to wormbase identifier
}
\details{
character vecotr with gene name or sequence name as names and
          wormbase identifier as values.
}
\examples{
data(ce.IDsMap)
head(ce.IDsMap)
}
\keyword{data}
