% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ce.interactionmap}
\alias{ce.interactionmap}
\title{transcript regulatory map of \emph{Caenorhabditis elegans}}
\format{datafram}
\source{
\url{http://edgedb.umassmed.edu}, \url{http://www.ebi.ac.uk/enright-srv/microcosm/htdocs/targets/v5/}
}
\usage{
ce.interactionmap
}
\description{
transcript regulatory map of \emph{Caenorhabditis elegans}
}
\details{
transcript regulatory map of \emph{Caenorhabditis elegans} is 
         generated using databases edgedb and microCosm Targets.
}
\examples{
data(ce.interactionmap)
head(ce.interactionmap)

}
\keyword{datasets}
