% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ce.mapIDs}
\alias{ce.mapIDs}
\title{map file for converting from wormbase identifier to \emph{Caenorhabditis elegans} gene name}
\format{character vector}
\source{
\url{http://www.wormbase.org/}
}
\usage{
ce.mapIDs
}
\description{
map file for converting from wormbase identifier to \emph{Caenorhabditis elegans} gene name
}
\details{
character vecotr with wormbase identifier as names and gene name as values.
}
\examples{
data(ce.mapIDs)
head(ce.mapIDs)

}
\keyword{datasets}
