% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ce.miRNA.map}
\alias{ce.miRNA.map}
\title{micro RNA of \emph{Caenorhabditis elegans}}
\format{dataframe}
\source{
\url{http://www.mirbase.org/}
}
\usage{
ce.miRNA.map
}
\description{
micro RNA of \emph{Caenorhabditis elegans}
}
\details{
The first column is wormbase identifier. And the second column is miRNA names.
}
\examples{
data(ce.miRNA.map)
head(ce.miRNA.map)

}
\keyword{datasets}

