\name{hs.IDsMap}
\alias{hs.IDsMap}
\docType{data}
\title{
  map file for converting gene name or sequence name of \emph{Homo sapiens} to Entrez identifier
}
\description{
  map file for converting gene name or sequence name of \emph{Homo sapiens} to Entrez identifier
}
\usage{data(hs.IDsMap)}
\format{
   character vector
}
\details{
	character vecotr with gene name as names and Entrez identifier as values.
}
\examples{
data(hs.IDsMap)
head(hs.IDsMap)
}
\keyword{datasets}
