% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{hs.interactionmap}
\alias{hs.interactionmap}
\title{transcript regulation map of \emph{Homo sapiens}}
\format{
dataframe
}
\source{
\url{http://fantom.gsc.riken.jp/5/}, 
        \url{http://www.ebi.ac.uk/enright-srv/microcosm/htdocs/targets/v5/},
        \url{http://carolina.imis.athena-innovation.gr/diana_tools/web/index.php}
}
\usage{
hs.interactionmap
}
\description{
transcript regulation map of \emph{Homo sapiens}
}
\details{
transcript regulatory map of \emph{Homo sapiens} is generated using 
         databases FANTOM, mirGen and microCosm Targets.
}
\examples{
data(hs.interactionmap)
head(hs.interactionmap)

}
\keyword{data}
