% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{hs.mapIDs}
\alias{hs.mapIDs}
\title{map file for converting from Entrez identifier to \emph{Homo sapiens} gene name}
\format{
character vector
}
\usage{
hs.mapIDs
}
\description{
map file for converting from Entrez identifier to \emph{Homo sapiens} gene name
}
\details{
character vecotr with Entrez identifier as names and gene name as values.
}
\examples{
data(hs.mapIDs)
head(hs.mapIDs)

}
\keyword{data}
