% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{hs.miRNA.map}
\alias{hs.miRNA.map}
\title{micro RNA of \emph{Homo sapiens}}
\format{dataframe}
\source{
\url{http://www.mirbase.org/}
}
\usage{
hs.miRNA.map
}
\description{
micro RNA of \emph{Homo sapiens}
}
\details{
The first column is entrez identifier. And the second column is miRNA names.
}
\examples{
data(hs.miRNA.map)
head(hs.miRNA.map)

}
\keyword{datasets}
