% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkBuilder.R
\name{uniqueExprsData}
\alias{uniqueExprsData}
\title{unique the microarray data}
\usage{
uniqueExprsData(exprsData, method = "Max", condenseName = "logFC")
}
\arguments{
\item{exprsData}{dataset of expression comparison data}

\item{method}{method must be Max, Median or Min}

\item{condenseName}{column names to be condensed}
}
\value{
a dataframe of expression data without duplicates
}
\description{
get unique the microarray data for each gene id.
}
\examples{
data("example.data")
example.microarrayData<-uniqueExprsData(example.data$example.microarrayData,
                                        method="Max", condenseName='logFC')
}
\keyword{network}

